requirements.txt
beautifulsoup4
pycryptodomex
requests
termcolor
tqdm
yarl
lxml